# feat(math): add attention output projection layer (#184)