## Test Coverage
- Current coverage: 87.5%
- Coverage changes: 87.3% → 87.5%

## Performance Metrics
### Memory Usage
#### Tensor Operations
- Allocations per operation:
  - New tensor creation: 120 allocs/op
  - Get/Set operations: 0 allocs/op
  - Parallel operations: 160750 allocs/op
  - BitLinear operations: 3290 allocs/op

#### BitNet Model Operations
- Allocations per operation:
  - Model weights loading: 1289985000 allocs/op
  - Model inference: N/A (TODO #190) allocs/op (TODO #190)
  - Ternary weights reading: 48 allocs/op

### CPU Performance
#### Tensor Operations
- Operation timing:
  - Basic operations: 11.80 ns/op
  - Parallel operations: 95489 ns/op
  - Large tensor operations: 1265 ns/op
  - BitLinear operations: 24563123 ns/op

#### BitNet Model Operations
- Operation timing:
  - Model weights loading: 1739052667 ns/op
  - Model inference: BenchmarkModel_Infer ns/op (TODO #190)
  - Ternary weights reading: 3814 ns/op

## Areas for Improvement
### High Priority
- [ ] Optimize memory allocations in model operations (TODO #191)
- [ ] Implement proper self-attention (TODO #186)

### Medium Priority
- [ ] Improve error handling in model operations (TODO #192)
- [ ] Add more comprehensive benchmarks (TODO #192)
- [ ] Enhance documentation
- [ ] Implement proper feed-forward network (TODO #187)

### Low Priority
- [ ] Consider SIMD optimizations (TODO #191)
- [ ] Add more model operations (TODO #190)
- [ ] Improve test organization (TODO #192)
- [ ] Implement proper output generation (TODO #189)

Closes #184